mecab: merge mecab-unidic as an optional resource
